+++ src/main/java/com/project/wah/projectwebservice/service/UsersService.java
@@ -6,8 +6,8 @@ import com.project.wah.projectwebservice.config.auth.dto.SessionUser;
 import com.project.wah.projectwebservice.domain.user.Role;
 import com.project.wah.projectwebservice.domain.user.User;
 import com.project.wah.projectwebservice.domain.user.UserRepository;
-import com.project.wah.projectwebservice.web.dto.UsersResponseDto;
-import com.project.wah.projectwebservice.web.dto.UsersUpdateRequestDto;
+import com.project.wah.projectwebservice.web.dto.user.UsersResponseDto;
+import com.project.wah.projectwebservice.web.dto.user.UsersUpdateRequestDto;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.stereotype.Service;
@@ -22,7 +22,7 @@ public class UsersService {
     private final UserRepository userRepository;
 
     @Transactional
-    public int update(SessionUser sessionUser, UsersUpdateRequestDto requestDto, HttpSession httpSession){
+    public Long update(SessionUser sessionUser, UsersUpdateRequestDto requestDto, HttpSession httpSession){
         User user = userRepository.findById(sessionUser.getId()).orElseThrow(() ->
                 new IllegalArgumentException("해당 유저가 없습니다. id="+ sessionUser.getId()));
 
@@ -35,7 +35,7 @@ public class UsersService {
     }
 
     @Transactional
-    public UsersResponseDto findById(int id) {
+    public UsersResponseDto findById(Long id) {
         User user = userRepository.findById(id).orElseThrow(() ->
                 new IllegalArgumentException("해당 유저가 없습니다. id="+ id));
 
@@ -43,7 +43,7 @@ public class UsersService {
     }
 
     @Transactional
-    public int delete (SessionUser sessionUser, HttpSession httpSession) {
+    public Long delete (SessionUser sessionUser, HttpSession httpSession) {
         User user = userRepository.findById(sessionUser.getId()).orElseThrow(() ->
                 new IllegalArgumentException("해당 유저가 없습니다. id="+ sessionUser.getId()));
